add download process bar 10
